LERMVP-40 Fixed the versioning.
TrainingWeightsResponse was not setting the version value, also the current weights should be version-1 from the task version. I did it in a not nice way, maybe we should keep weights version and task version separately or have task version to be weights+1 or just leave it as it is not

I changed the way jobs are loaded by introducing the ApplicationStartup class. Don't kill me for the atrocities I'm doing in there and the function I made public from the FLManager. [LERMVP-28]

Everything works for multiple simultaneous devices though!

Fixup! Moved the startup code to flmanager and changed the version on the app table to be the current weights version and the task version is the weights version+1
Fixup! Fixed the case where the getGlobalWeights always returned the weights no matter the version of the mobile weights, the first time the getGlobalWeights was running